#233 Refactor bulk editor
* Remove BulkEditSaveModal - not used
* Move grid renderer components
* Rename base component to `bulk-edit-container`
