+++ src/app/bulk-edit/bulk-edit-editor/renderers/checkbox-renderer/checkbox-renderer.component.ts
