--- src/app/bulk-edit/modals/bulk-edit-save-modal/bulk-edit-save-modal.component.html
@@ -1,1 +1,0 @@
-<p>bulk-edit-save-modal works!</p>
