--- src/app/bulk-edit/modals/bulk-edit-save-modal/bulk-edit-save-modal.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'mdm-bulk-edit-save-modal',
-  templateUrl: './bulk-edit-save-modal.component.html',
-  styleUrls: ['./bulk-edit-save-modal.component.scss']
-})
-export class BulkEditSaveModalComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
